Updated application for Laravel 5.3
