--- app/Http/Controllers/Auth/PasswordController.php
@@ -1,87 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\User;
-use Illuminate\Http\Request;
-use Illuminate\Mail\Message;
-use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Password;
-use Illuminate\Foundation\Auth\ResetsPasswords;
-
-class PasswordController extends Controller
-{
-    /*
-    |--------------------------------------------------------------------------
-    | Password Reset Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller is responsible for handling password reset requests
-    | and uses a simple trait to include this behavior. You're free to
-    | explore this trait and override any methods you wish to tweak.
-    |
-    */
-
-    use ResetsPasswords;
-    protected $redirectTo = '/';
-
-    /**
-     * Send a reset link to the given user.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function sendResetLinkEmail(Request $request)
-    {
-        $this->validateSendResetLinkEmail($request);
-        $broker = $this->getBroker();
-        $response = Password::broker($broker)->sendResetLink(
-            $this->getSendResetLinkEmailCredentials($request),
-            $this->resetEmailBuilder(User::where('email', $request->email)->firstOrFail())
-        );
-
-
-
-        switch ($response) {
-            case Password::RESET_LINK_SENT:
-                return $this->getSendResetLinkEmailSuccessResponse($response);
-            case Password::INVALID_USER:
-            default:
-                return $this->getSendResetLinkEmailFailureResponse($response);
-        }
-    }
-
-
-    /**
-     * Get the Closure which is used to build the password reset email message.
-     *
-     * @return \Closure
-     */
-    protected function resetEmailBuilder($name)
-    {
-        return function (Message $message, $name) {
-            $message->subject($this->getEmailSubject($name));
-        };
-    }
-
-    /**
-     * Get the e-mail subject line to be used for the reset link email.
-     *
-     * @return string
-     */
-    protected function getEmailSubject($name)
-    {
-        return '['.$name->name.'] Your Password Reset Link';
-    }
-
-
-    /**
-     * Create a new password controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware($this->guestMiddleware());
-    }
-}
