+++ app/Http/Controllers/Auth/RegisterController.php
@@ -27,7 +27,7 @@ class RegisterController extends Controller
      *
      * @var string
      */
-    protected $redirectTo = '/home';
+    protected $redirectTo = '/';
 
     /**
      * Create a new controller instance.
@@ -67,5 +67,15 @@ class RegisterController extends Controller
             'email' => $data['email'],
             'password' => bcrypt($data['password']),
         ]);
+    }
+
+    public function register()
+    {
+        return redirect('/');
+    }
+
+    public function showRegistrationForm()
+    {
+        return redirect('/');
     }
 }
