+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -2,7 +2,11 @@
 
 namespace App\Http\Controllers\Auth;
 
+use App\User;
+use Illuminate\Http\Request;
+use Illuminate\Mail\Message;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Password;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 
 class ResetPasswordController extends Controller
@@ -19,21 +23,65 @@ class ResetPasswordController extends Controller
     */
 
     use ResetsPasswords;
+    protected $redirectTo = '/';
 
     /**
-     * Where to redirect users after resetting their password.
+     * Send a reset link to the given user.
      *
-     * @var string
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
      */
-    protected $redirectTo = '/home';
+    public function sendResetLinkEmail(Request $request)
+    {
+        $this->validateSendResetLinkEmail($request);
+        $broker = $this->getBroker();
+        $response = Password::broker($broker)->sendResetLink(
+            $this->getSendResetLinkEmailCredentials($request),
+            $this->resetEmailBuilder(User::where('email', $request->email)->firstOrFail())
+        );
+
+
+
+        switch ($response) {
+            case Password::RESET_LINK_SENT:
+                return $this->getSendResetLinkEmailSuccessResponse($response);
+            case Password::INVALID_USER:
+            default:
+                return $this->getSendResetLinkEmailFailureResponse($response);
+        }
+    }
+
 
     /**
-     * Create a new controller instance.
+     * Get the Closure which is used to build the password reset email message.
+     *
+     * @return \Closure
+     */
+    protected function resetEmailBuilder($name)
+    {
+        return function (Message $message, $name) {
+            $message->subject($this->getEmailSubject($name));
+        };
+    }
+
+    /**
+     * Get the e-mail subject line to be used for the reset link email.
+     *
+     * @return string
+     */
+    protected function getEmailSubject($name)
+    {
+        return '['.$name->name.'] Your Password Reset Link';
+    }
+
+
+    /**
+     * Create a new password controller instance.
      *
      * @return void
      */
     public function __construct()
     {
-        $this->middleware('guest');
+        $this->middleware($this->guestMiddleware());
     }
 }
