--- app/Providers/PaginationServiceProvider.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace App\Providers;
-
-use Illuminate\Pagination\Paginator;
-use Illuminate\Support\ServiceProvider;
-use Illuminate\Pagination\BootstrapFourPresenter;
-
-class PaginationServiceProvider extends ServiceProvider
-{
-    /**
-     * Bootstrap the application services.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        //
-    }
-
-    /**
-     * Register the application services.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        Paginator::presenter(function ($paginator) {
-            return new BootstrapFourPresenter($paginator);
-        });
-    }
-}
