+++ readme.md
@@ -17,12 +17,13 @@ Develop: [![Build Status](https://travis-ci.org/shivampaw/UltimateCRM.svg?branch
 ## Requirements / Installation / Troubleshooting
 Need help getting started? Here's where to look!
 ### Requirements
-This CRM runs with Laravel 5.2 for which the requirements are as follows:
-* PHP >= 5.5.9
+This CRM runs with Laravel 5.3 for which the requirements are as follows:
+* PHP >= 5.6.4
 * OpenSSL PHP Extension
 * PDO PHP Extension
 * Mbstring PHP Extension
 * Tokenizer PHP Extension
+* XML PHP Extension
 
 ### Installation
 To install run the following commands in a working directory:
@@ -32,7 +33,7 @@ You will then need to change your web server root to the public directory. Conta
 
 Now run `composer install` and make sure you have composer installed on your server!
 
-Then setup your database and open the `.env.example` file. Rename this file to `.env` and enter details for your database, stripe and email integration.
+Then setup your database and open the `.env.example` file. Rename this file to `.env` and enter details for your site, database, stripe and email integration.
 
 Run `php artisan key:generate` from the root project folder.
